refactor(service): 도메인 로직 분리를 위한 service 패키지 위치 변경
- LottoConfig 및 Controller 내 import 경로 수정
-  로직 변경 없이 계층 구조를 명확히 정리
