+++ src/main/java/lotto/config/LottoConfig.java
@@ -6,7 +6,7 @@ import lotto.domain.factory.LottoNumberFactory;
 import lotto.domain.factory.PurchaseAmountFactory;
 import lotto.domain.generator.LottoNumberGenerator;
 import lotto.domain.generator.RandomLottoNumberGenerator;
-import lotto.domain.service.LottoService;
+import lotto.service.LottoService;
 import lotto.util.validator.LottoNumberValidator;
 import lotto.util.validator.LottoValidator;
 import lotto.util.validator.NumberValidator;
